golang struct representations with json annotations that are directly compatible with the CUE schema
I tried to get CUE to generate the go structs directly, it seems that functionality is still in beta but rather than dig into that I hand-rolled the equivalent schema.  Supposing I make enough progress on the client side of Q-Party I will venture into that CUE -> Go conversion and see why it won't export to code/go filetype.. or at least collect enough information to make a solid bug report

for now, though, this will serve as the type that is used when sending JSON to the client, and will serve well as an intermediate type during episode archival and editing.  The schema will change some but not suddenly.