refactor html-parsing functionality to its own module
This is also a step towards removing any data structures that are only for transitively holding the contents of the HTML documents.  They were useful for getting the parsing routines working but if this code is going to stick around it should just be what's needed to read an individual episode's questions for simulating.

the next CL will dispose of any structs that still have a `json...` tag attached to any properties, those are all either deprecated or belong in the project's encoding/json directory