the episode's Show Number is not actually unique, it is only unique to regular season showings
the special tournaments and other shows that were not part of a regular season all get their own 1..n numeric range, so we are probably seeing multiple collisions in the lower dozen or so numbers.

this change makes EpisodeID a canonical identifier, and ShowNumber is no longer a uint, it is now a string value which may have an alphabetic prefix or be prefixed with the '#' (octothorpe) symbol.